+++ src/mainej/headlessui_reagent.cljs
@@ -33,6 +33,11 @@
 (def popover-panel (utils/headlessui->reagent ui/Popover.Panel))
 (def popover-group (utils/headlessui->reagent ui/Popover.Group))
 
+(def radio-group (utils/headlessui->reagent ui/RadioGroup))
+(def radio-group-option (utils/headlessui->reagent ui/RadioGroup.Option))
+(def radio-group-label (utils/headlessui->reagent ui/RadioGroup.Label))
+(def radio-group-description (utils/headlessui->reagent ui/RadioGroup.Description))
+
 (defn transition [& args] (into [:> ui/Transition] args))
 (defn transition-child [& args] (into [:> ui/Transition.Child] args))